update tests for physical database, page cache, page, header page, node, relationship and heap file to be compliant to system catalogue change

+++ include/access/header_page.h
@@ -37,14 +37,14 @@ read_bits(page_cache*    pc,
           page*          p,
           unsigned short byte_offset_in_page,
           unsigned char  bit_offset_in_byte,
-          unsigned short n_bits);
+          unsigned long  n_bits);
 
 void
 write_bits(page_cache*    pc,
            page*          p,
            unsigned short byte_offset_in_page,
            unsigned char  bit_offset_in_byte,
-           unsigned short n_bits,
+           unsigned long  n_bits,
            unsigned char* data);
 
 #endif
